Generated DropDowns
This change generates the menue points of the DropDown from the database
entries in sorted order. I factored it into one method, the
generateDropDown(String,ListBox) method, and just call that method in
onModuleLoad. Dont forget the imports. The util.Arrays, util.Collections
and util.HashSet are crucial for this method to work
Sweet thanks have fun :dancer:  :D :+1:  :100: